backend/requirements.txt
django
djangorestframework
django-cors-headers
pre-commit
